boilerplate

+++ boilerplate/update.sh
@@ -1,23 +1,21 @@
 #!/usr/bin/env bash
 
 # WARNING: THIS FILE IS MANAGED IN THE 'BOILERPLATE' REPO AND COPIED TO OTHER REPOSITORIES.
-# ONLY EDIT THIS FILE FROM WITHIN THE 'LYFT/BOILERPLATE' REPOSITORY:
+# ONLY EDIT THIS FILE FROM WITHIN THE 'FLYTEORG/BOILERPLATE' REPOSITORY:
 # 
-# TO OPT OUT OF UPDATES, SEE https://github.com/lyft/boilerplate/blob/master/Readme.rst
+# TO OPT OUT OF UPDATES, SEE https://github.com/flyteorg/boilerplate/blob/master/Readme.rst
 
 set -e
 
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
 
 OUT="$(mktemp -d)"
-trap "rm -fr $OUT" EXIT
+trap 'rm -fr $OUT' EXIT
 
-git clone git@github.com:flyteorg/boilerplate.git "${OUT}"
+git clone https://github.com/flyteorg/boilerplate.git "${OUT}" 
 
 echo "Updating the update.sh script."
 cp "${OUT}/boilerplate/update.sh" "${DIR}/update.sh"
-echo ""
-
 
 CONFIG_FILE="${DIR}/update.cfg"
 README="https://github.com/flyteorg/boilerplate/blob/master/Readme.rst"
@@ -30,12 +28,12 @@ if [ ! -f "$CONFIG_FILE" ]; then
 fi
 
 if [ -z "$REPOSITORY" ]; then
-  echo '$REPOSITORY is required to run this script'
+  echo "$REPOSITORY is required to run this script"
   echo "See $README for more details."
   exit 1
 fi
 
-while read directory junk; do
+while read -r directory junk; do
   # Skip comment lines (which can have leading whitespace)
   if [[ "$directory" == '#'* ]]; then
     continue
@@ -63,8 +61,8 @@ while read directory junk; do
   echo "$directory is configured in update.cfg."
   echo "-----------------------------------------------------------------------------------"
   echo "syncing files from source."
-  rm -rf "${DIR}/${directory}"
-  mkdir -p $(dirname "${DIR}/${directory}")
+  rm -rf "${DIR:?}/${directory}"
+  mkdir -p "$(dirname "${DIR}"/"${directory}")"
   cp -r "$dir_path" "${DIR}/${directory}"
   if [ -f "${DIR}/${directory}/update.sh" ]; then
     echo "executing ${DIR}/${directory}/update.sh"
